Fix linker flags error

+++ xcodeNative/IASP/IASP/NoticiasViewController.m
@@ -150,7 +150,7 @@
 #
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSLog(@"index %i", indexPath.row);
+    NSLog(@"index %li", (long)indexPath.row);
     
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     [self performSegueWithIdentifier:@"noticiasDetails" sender:indexPath];
